[REFACTOR] Re-orginzed Profile files
