+++ src/App.js
@@ -1,18 +1,18 @@
 import { useContext } from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import Layout from './components/Layout/Layout';
-import UserProfile from './components/Profile/UserProfile';
+import UserProfile from './pages/UserProfile/UserProfile';
 import AuthPage from './pages/AuthPage';
 import HomePage from './pages/HomePage';
 import AuthContext from './store/auth-context';
 import Loggedin from './pages/Loggedin';
 import AddProduct from './components/Profile/AddProduct';
-import Inventory from './components/Profile/Inventory';
+import Inventory from './pages/ProductDetails/Inventory';
 import EmployeeDisplay from './pages/EmployeeDisplay';
-import RevenueReport from './components/Profile/RevenueReport';
-import JobCard from './components/Profile/JobCard';
+import RevenueReport from './pages/RevenueReport/RevenueReport';
+import JobCard from './pages/JobCard/JobCard';
 import EstimationPage from './pages/EstimationPage';
-import Invoice from './components/Profile/invoice';
+import Invoice from './pages/BillGeneration/invoice';
 import SheetDisplay from './components/EmployeeDetails/EmployeeDisplay/SheetDisplay';
