+++ src/pages/JobCard/JobCard.css
